backend/requirements.txt
amqp==2.4.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
autobahn==19.7.2
Automat==0.7.0
beautifulsoup4==4.6.3
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.1.4
django-cors-headers==2.5.2
django-dynamic-fixture==2.0.0
django-environ==0.4.5
django-webpack-loader==0.6.0
djangorestframework==3.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
kombu==4.3.0
lxml==4.3.0
more-itertools==7.0.0
parsel==1.5.1
pem==19.1.0
Pillow==5.3.0
pluggy==0.9.0
psutil==5.6.3
psycopg2==2.8.2
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pytest==4.4.1
pytest-django==3.4.8
python-dateutil==2.8.0
pytz==2018.7
queuelib==1.5.0
requests==2.21.0
schedule==0.6.0
Scrapy==1.5.2
scrapyd==1.2.1
scrapyd-client==1.2.0a1
service-identity==18.1.0
six==1.12.0
tabulate==0.8.3
Twisted==18.9.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.24.2
vine==1.2.0
w3lib==1.20.0
zope.interface==4.6.0
